# docs/swagger.yaml
basePath: /
definitions:
...
  controller._ResponseMenuList:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/controller.ResCode'
        description: 业务响应状态码
      data:
        $ref: '#/definitions/model.Menu'
      message:
        description: 提示信息
        type: string
      ok:
        type: boolean
    type: object
...
  controller._ResponseTmList:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/controller.ResCode'
        description: 业务响应状态码
      data:
        $ref: '#/definitions/model.ResponseTmList'
      message:
        description: 提示信息
        type: string
      ok:
        type: boolean
    type: object
  controller._ResponseToAssignMenuList:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/controller.ResCode'
        description: 业务响应状态码
      data:
        items:
          $ref: '#/definitions/model.Menu'
        type: array
      message:
        description: 提示信息
        type: string
      ok:
        type: boolean
    type: object
...
  controller.ResCode:
    enum:
    - 200
    - 201
    - 202
    - 203
    - 204
    - 205
    - 206
    - 207
    - 208
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeInvalidParam
    - CodeUserExist
    - CodeUserNotExist
    - CodeInvalidPassword
    - CodeServerBusy
    - CodeInvalidToken
    - CodeNeedLogin
    - CodeMenuNodeExist
  controller.ResponseData:
    properties:
      code:
        $ref: '#/definitions/controller.ResCode'
      data: {}
      message: {}
      ok:
        type: boolean
    type: object
  model.Attr:
    properties:
      ID:
        type: integer
      attrName:
        description: 属性名称
        type: string
      attrValueList:
        description: 属性值列表
        items:
          $ref: '#/definitions/model.AttrValue'
        type: array
      categoryId:
        description: 三级分类ID
        type: integer
      categoryLevel:
        description: 分类级别
        type: integer
      createTime:
        type: string
      id:
        description: 属性ID
        type: integer
      updateTime:
        type: string
    required:
    - attrName
    - categoryId
    - categoryLevel
    type: object
  model.AttrValue:
    properties:
      attrId:
        type: integer
      id:
        type: integer
      valueName:
        type: string
    required:
    - valueName
    type: object
  model.Menu:
    properties:
      ID:
        type: integer
      children:
        items:
          $ref: '#/definitions/model.Menu'
        type: array
      code:
        type: string
      createTime:
        type: string
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      pid:
        type: integer
      select:
        type: boolean
      status:
        type: string
      toCode:
        type: string
      type:
        type: integer
      updateTime:
        type: string
    type: object
...
  model.ParamMenuSave:
    properties:
      code:
        type: string
      level:
        type: integer
      name:
        type: string
      pid:
        type: integer
      type:
        type: integer
    type: object
  model.ParamMenuUpdate:
    properties:
      code:
        type: string
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      pid:
        type: integer
    type: object
...
  model.ResponseTmList:
    properties:
      current:
        type: integer
      pages:
        type: integer
      records:
        items:
          $ref: '#/definitions/model.Trademark'
        type: array
      searchCount:
        type: boolean
      size:
        type: integer
      total:
        type: integer
    type: object
...
  model.Trademark:
    properties:
      ID:
        type: integer
      createTime:
        type: string
      id:
        type: integer
      logoUrl:
        type: string
      tmName:
        type: string
      updateTime:
        type: string
    type: object
host: 127.0.0.1:10086
info:
  contact:
    email: 929000201@qq.com
    name: WangJian
    url: https://github.com/wangjian3306
  description: 硅谷甄选项目后端
  title: vue3_admin 项目接口文档
  version: "1.0"
paths:
...
  /admin/acl/permission:
    get:
      consumes:
      - application/json
      description: 获取菜单列表
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseMenuList'
      security:
      - ApiKeyAuth: []
      summary: 获取菜单列表
      tags:
      - 菜单管理
  /admin/acl/permission/doAssign:
    post:
      consumes:
      - application/json
      description: 给角色分配权限
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 角色 ID
        in: query
        name: roleId
        required: true
        type: integer
      - collectionFormat: csv
        description: 菜单 ID列表
        in: query
        items:
          type: integer
        name: permissionId
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 给角色分配权限
      tags:
      - 菜单管理
  /admin/acl/permission/remove/{id}:
    delete:
      consumes:
      - application/json
      description: 处理删除菜单请求
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 菜单 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 删除菜单接口
      tags:
      - 菜单管理
  /admin/acl/permission/save:
    post:
      consumes:
      - application/json
      description: 处理新增菜单请求
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 用户登录参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ParamMenuSave'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 新增菜单接口
      tags:
      - 菜单管理
  /admin/acl/permission/toAssign/{roleId}:
    get:
      consumes:
      - application/json
      description: 根据角色获取菜单请求
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 角色ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseToAssignMenuList'
      security:
      - ApiKeyAuth: []
      summary: 根据角色获取菜单接口
      tags:
      - 菜单管理
  /admin/acl/permission/update:
    put:
      consumes:
      - application/json
      description: 处理更新菜单请求
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 用户登录参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ParamMenuUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 更新菜单接口
      tags:
      - 菜单管理
...
  /admin/deleteAttr/{attrId}:
    delete:
      consumes:
      - application/json
      description: 处理删除基础属性请求
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 属性 ID
        in: path
        name: attrId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 删除基础属性接口
      tags:
      - 商品基础属性接口
  /admin/product/attrInfoList/{c1Id}/{c2Id}/{c3Id}:
    get:
      consumes:
      - application/json
      description: 处理获取分类下已有的属性与属性值请求
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 分类一 ID
        in: path
        name: c1Id
        required: true
        type: integer
      - description: 分类二 ID
        in: path
        name: c2Id
        required: true
        type: integer
      - description: 分类三 ID
        in: path
        name: c3Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseToAssignMenuList'
      security:
      - ApiKeyAuth: []
      summary: 获取分类下已有的属性与属性值接口
      tags:
      - 商品基础属性接口
...
  /admin/product/fileUpload:
    post:
      consumes:
      - application/json
      description: 上传文件
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 上传文件接口
      tags:
      - 上传文件
  /admin/product/getCategory1:
    get:
      consumes:
      - application/json
      description: 获取一级分类
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseTmList'
      security:
      - ApiKeyAuth: []
      summary: 获取一级分类接口
      tags:
      - 商品分类接口
  /admin/product/getCategory2/{id}:
    get:
      consumes:
      - application/json
      description: 获取二级分类
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 分类一 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseTmList'
      security:
      - ApiKeyAuth: []
      summary: 获取二级分类接口
      tags:
      - 商品分类接口
  /admin/product/getCategory3/{id}:
    get:
      consumes:
      - application/json
      description: 获取三级分类
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 分类二 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseTmList'
      security:
      - ApiKeyAuth: []
      summary: 获取三级分类接口
      tags:
      - 商品分类接口
  /admin/product/saveAttrInfo:
    post:
      consumes:
      - application/json
      description: 处理添加或者修改已有的属性请求
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 属性信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.Attr'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 添加或者修改已有的属性的接口
      tags:
      - 商品基础属性接口
swagger: "2.0"